AdventOfCodeTests/Day04Tests.cs
using System;
using Microsoft.VisualStudio.TestTools.UnitTesting;

namespace AdventOfCodeTests
{
    [TestClass]
    public class Day04Tests
    {
        private string input_puzzle;
        private string input_example1;
        private string input_example2;

        [TestInitialize]
        public void LoadInput()
        {
            string day = "04";
            input_puzzle = Resources.Input.ResourceManager.GetObject($"D{day}_Puzzle").ToString();
            input_example1 = string.Format("ecl:gry pid:860033327 eyr:2020 hcl:#fffffd{0}byr:1937 iyr:2017 cid:147 hgt:183cm{0}{0}iyr:2013 ecl:amb cid:350 eyr:2023 pid:028048884{0}hcl:#cfa07d byr:1929{0}{0}hcl:#ae17e1 iyr:2013{0}eyr:2024{0}ecl:brn pid:760753108 byr:1931{0}hgt:179cm{0}{0}hcl:#cfa07d eyr:2025 pid:166559648{0}iyr:2011 ecl:brn hgt:59in", Environment.NewLine);
            input_example2 = string.Format("example{0}2", Environment.NewLine);
        }

        [TestMethod]
        public void Puzzle0_WarmUp()
        {
            // Force performing LoadInput() warm-up as part of this test
        }

        [TestMethod]
        public void Puzzle1()
        {
            // Act
            var result = AdventOfCode.Day04.Puzzle1(input_puzzle);

            // Assert
            Assert.AreEqual("210", result);
        }

        [TestMethod]
        public void Puzzle1_Example()
        {
            // Act
            var result = AdventOfCode.Day04.Puzzle1(input_example1);

            // Assert
            Assert.AreEqual("2", result);
        }

        [TestMethod]
        public void Puzzle2()
        {
            // Act
            var result = AdventOfCode.Day04.Puzzle2(input_puzzle);

            // Assert
            Assert.AreEqual($"{input_puzzle}_Puzzle2", result);
        }

        [TestMethod]
        public void Puzzle2_Example()
        {
            // Act
            var result = AdventOfCode.Day04.Puzzle2(input_example2);

            // Assert
            Assert.AreEqual($"{input_example2}_Puzzle2", result);
        }
    }
}
